Rename internal to server.
