+++ api/saywhat/router.go
@@ -9,7 +9,7 @@ package saywhat
 import (
 	"github.com/gin-gonic/gin"
 
-	"clickonetwo.io/whisper/internal/middleware"
+	"clickonetwo.io/whisper/server/middleware"
 )
 
 func AddRoutes(r *gin.RouterGroup) {
