+++ cmd/utils.go
@@ -12,10 +12,10 @@ import (
 	"os"
 	"strings"
 
-	"clickonetwo.io/whisper/internal/client"
-	"clickonetwo.io/whisper/internal/conversation"
-	"clickonetwo.io/whisper/internal/profile"
-	"clickonetwo.io/whisper/internal/storage"
+	"clickonetwo.io/whisper/server/client"
+	"clickonetwo.io/whisper/server/conversation"
+	"clickonetwo.io/whisper/server/profile"
+	"clickonetwo.io/whisper/server/storage"
 )
 
 func saveObjects(what storage.ObjectMap) {
