+++ server/client/storage.go
@@ -9,7 +9,7 @@ package client
 import (
 	"fmt"
 
-	"clickonetwo.io/whisper/internal/storage"
+	"clickonetwo.io/whisper/server/storage"
 )
 
 type Data struct {
