+++ server/conversation/storage.go
@@ -9,7 +9,7 @@ package conversation
 import (
 	"fmt"
 
-	"clickonetwo.io/whisper/internal/storage"
+	"clickonetwo.io/whisper/server/storage"
 )
 
 type Data struct {
